requirements.txt
# FastAPI e dependências web
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Banco de dados
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Google Cloud Storage
google-cloud-storage==2.10.0

# Parsing de planilhas (mantendo as mesmas bibliotecas do script original)
pandas==2.1.3
numpy==1.25.2
ezodf==0.3.2
openpyxl==3.1.2
xlrd==2.0.1

# Validação e serialização
pydantic[email]==2.5.0
pydantic-settings==2.0.3

# Utilitários
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.25.2